Moved prediction serializer
