+++ src/mlmodule/cli/run.py
@@ -25,7 +25,7 @@ from mlmodule.cli.types import (
     CLIObjectFactory,
 )
 from mlmodule.predictions import BatchModelPrediction
-from mlmodule.serializers import batch_model_prediction_to_dict
+from mlmodule.predictions.serializers import batch_model_prediction_to_dict
 from mlmodule.v2.states import StateKey
 from mlmodule.v2.stores import Store
 from mlmodule.v2.stores.abstract import AbstractStateStore
